REF: drop dataRevision; refresh via setWallets without React.memo
Glados asked to fix stale cards at the source instead of revision-string
algebra. Keep publishing wallets after in-place fetch and SectionList
extraData; remove React.memo/dataRevision so cards re-render when that
publish lands. Retain shared walletHasPendingTransaction helper + tests.
