+++ components/WalletsCarousel.tsx
@@ -74,17 +74,6 @@ export const walletHasPendingTransaction = (item: TWallet): boolean => {
   return item.getTransactions().some((tx: Transaction) => tx.confirmations === 0);
 };
 
-/** Primitive snapshot so React.memo notices in-place wallet mutations after refresh. */
-export const getWalletCarouselItemDataRevision = (item: TWallet): string =>
-  [
-    item.getBalance(),
-    item.getLatestTransactionTime(),
-    item.getLabel?.() ?? '',
-    item.hideBalance ? 1 : 0,
-    walletHasPendingTransaction(item) ? 1 : 0,
-    item.getPreferredBalanceUnit?.() ?? '',
-  ].join('|');
-
 interface NewWalletPanelProps {
   onPress: () => void;
 }
@@ -181,8 +170,6 @@ const NewWalletPanel: React.FC<NewWalletPanelProps> = ({ onPress }) => {
 interface WalletCarouselItemProps {
   item: TWallet;
   hideBalance: boolean;
-  /** See `getWalletCarouselItemDataRevision` — busts React.memo on in-place wallet updates. */
-  dataRevision: string;
   onPress: (item: TWallet) => void;
   handleLongPress?: () => void;
   isSelectedWallet?: boolean;
@@ -302,260 +289,252 @@ const iStyles = StyleSheet.create({
   },
 });
 
-export const WalletCarouselItem: React.FC<WalletCarouselItemProps> = React.memo(
-  ({
-    item,
-    hideBalance,
-    dataRevision,
-    onPress,
-    handleLongPress,
-    isSelectedWallet,
-    customStyle,
-    horizontal,
-    searchQuery,
-    renderHighlightedText,
-    animationsEnabled = true,
-    isPlaceHolder = false,
-    onPressIn,
-    onPressOut,
-    isNewWallet = false,
-    isExiting = false,
-    isDraggingActive = false,
-    dragActiveScale = 1.02,
-    sizeVariant = 'default',
-  }: WalletCarouselItemProps) => {
-    const walletLabel = item.getLabel ? item.getLabel() : '';
-    const pressScale = useSharedValue(1.0);
-    const dragScale = useSharedValue(isDraggingActive ? dragActiveScale : 1.0);
-    const opacityValue = useSharedValue(isSelectedWallet === false ? 0.5 : 1.0);
-    const translateYValue = useSharedValue(isNewWallet ? 20 : 0);
-    const balanceOpacity = useSharedValue(1);
-    const balanceTranslateY = useSharedValue(0);
-    const { colors } = useTheme();
-    const { width, fontScale } = useWindowDimensions();
-    const itemWidth = getWalletCarouselItemWidth(width);
-    const { sizeClass } = useSizeClass();
-    const isCompact = sizeVariant === 'compact';
-    const { direction } = useLocale();
-    const scaledCardStyles = useMemo(
-      () => ({
-        grad: { minHeight: getWalletCardMinHeight(fontScale) },
-        gradContent: { padding: scaleLayoutUp(15, fontScale) },
-        balanceContainer: { minHeight: scaleLayoutUp(40, fontScale) },
-        textSpacer: { height: scaleLayoutUp(WALLET_CARD_SECTION_GAP, fontScale) },
-        label: { lineHeight: scaleLayoutUp(24, fontScale) },
-        balance: { lineHeight: scaleLayoutUp(38, fontScale) },
-        balanceCompact: { lineHeight: scaleLayoutUp(30, fontScale) },
-        latestTx: { lineHeight: scaleLayoutUp(18, fontScale) },
-        latestTxTime: { lineHeight: scaleLayoutUp(22, fontScale) },
-      }),
-      [fontScale],
-    );
-    const cardTextStyle = useMemo(
-      () => ({
-        color: withAlpha(colors.inverseForegroundColor, WALLET_CARD_TEXT_OPACITY),
-        writingDirection: direction,
-      }),
-      [colors.inverseForegroundColor, direction],
-    );
-    const previousBalance = useRef<string | undefined>(undefined);
-    const balance = !hideBalance && formatBalance(Number(item.getBalance()), item.getPreferredBalanceUnit(), true);
-    const safeBalance = balance || undefined;
+export const WalletCarouselItem: React.FC<WalletCarouselItemProps> = ({
+  item,
+  hideBalance,
+  onPress,
+  handleLongPress,
+  isSelectedWallet,
+  customStyle,
+  horizontal,
+  searchQuery,
+  renderHighlightedText,
+  animationsEnabled = true,
+  isPlaceHolder = false,
+  onPressIn,
+  onPressOut,
+  isNewWallet = false,
+  isExiting = false,
+  isDraggingActive = false,
+  dragActiveScale = 1.02,
+  sizeVariant = 'default',
+}) => {
+  const walletLabel = item.getLabel ? item.getLabel() : '';
+  const pressScale = useSharedValue(1.0);
+  const dragScale = useSharedValue(isDraggingActive ? dragActiveScale : 1.0);
+  const opacityValue = useSharedValue(isSelectedWallet === false ? 0.5 : 1.0);
+  const translateYValue = useSharedValue(isNewWallet ? 20 : 0);
+  const balanceOpacity = useSharedValue(1);
+  const balanceTranslateY = useSharedValue(0);
+  const { colors } = useTheme();
+  const { width, fontScale } = useWindowDimensions();
+  const itemWidth = getWalletCarouselItemWidth(width);
+  const { sizeClass } = useSizeClass();
+  const isCompact = sizeVariant === 'compact';
+  const { direction } = useLocale();
+  const scaledCardStyles = useMemo(
+    () => ({
+      grad: { minHeight: getWalletCardMinHeight(fontScale) },
+      gradContent: { padding: scaleLayoutUp(15, fontScale) },
+      balanceContainer: { minHeight: scaleLayoutUp(40, fontScale) },
+      textSpacer: { height: scaleLayoutUp(WALLET_CARD_SECTION_GAP, fontScale) },
+      label: { lineHeight: scaleLayoutUp(24, fontScale) },
+      balance: { lineHeight: scaleLayoutUp(38, fontScale) },
+      balanceCompact: { lineHeight: scaleLayoutUp(30, fontScale) },
+      latestTx: { lineHeight: scaleLayoutUp(18, fontScale) },
+      latestTxTime: { lineHeight: scaleLayoutUp(22, fontScale) },
+    }),
+    [fontScale],
+  );
+  const cardTextStyle = useMemo(
+    () => ({
+      color: withAlpha(colors.inverseForegroundColor, WALLET_CARD_TEXT_OPACITY),
+      writingDirection: direction,
+    }),
+    [colors.inverseForegroundColor, direction],
+  );
+  const previousBalance = useRef<string | undefined>(undefined);
+  const balance = !hideBalance && formatBalance(Number(item.getBalance()), item.getPreferredBalanceUnit(), true);
+  const safeBalance = balance || undefined;
 
-    const animatePressScale = useCallback(
-      (toValue: number) => {
-        pressScale.value = withSpring(toValue, { damping: 13, stiffness: 180, mass: 0.9 });
-      },
-      [pressScale],
-    );
+  const animatePressScale = useCallback(
+    (toValue: number) => {
+      pressScale.value = withSpring(toValue, { damping: 13, stiffness: 180, mass: 0.9 });
+    },
+    [pressScale],
+  );
 
-    useEffect(() => {
-      dragScale.value = withSpring(isDraggingActive ? dragActiveScale : 1, { damping: 16, stiffness: 200, mass: 1 });
-    }, [isDraggingActive, dragActiveScale, dragScale]);
+  useEffect(() => {
+    dragScale.value = withSpring(isDraggingActive ? dragActiveScale : 1, { damping: 16, stiffness: 200, mass: 1 });
+  }, [isDraggingActive, dragActiveScale, dragScale]);
 
-    useEffect(() => {
-      if (!animationsEnabled) return;
+  useEffect(() => {
+    if (!animationsEnabled) return;
 
-      const targetOpacity = isSelectedWallet === false ? 0.5 : 1.0;
-      opacityValue.value = withSpring(targetOpacity, { damping: 18, stiffness: 240 });
-    }, [isSelectedWallet, opacityValue, animationsEnabled]);
+    const targetOpacity = isSelectedWallet === false ? 0.5 : 1.0;
+    opacityValue.value = withSpring(targetOpacity, { damping: 18, stiffness: 240 });
+  }, [isSelectedWallet, opacityValue, animationsEnabled]);
 
-    const onPressedIn = useCallback(() => {
-      if (animationsEnabled) {
-        animatePressScale(0.97);
-      }
-      if (onPressIn) onPressIn();
-    }, [animatePressScale, animationsEnabled, onPressIn]);
+  const onPressedIn = useCallback(() => {
+    if (animationsEnabled) {
+      animatePressScale(0.97);
+    }
+    if (onPressIn) onPressIn();
+  }, [animatePressScale, animationsEnabled, onPressIn]);
 
-    const onPressedOut = useCallback(() => {
-      if (animationsEnabled) {
-        animatePressScale(1.0);
-      }
-      if (onPressOut) onPressOut();
-    }, [animatePressScale, animationsEnabled, onPressOut]);
+  const onPressedOut = useCallback(() => {
+    if (animationsEnabled) {
+      animatePressScale(1.0);
+    }
+    if (onPressOut) onPressOut();
+  }, [animatePressScale, animationsEnabled, onPressOut]);
 
-    const handlePress = useCallback(() => {
-      onPress(item);
-    }, [item, onPress]);
+  const handlePress = useCallback(() => {
+    onPress(item);
+  }, [item, onPress]);
 
-    useEffect(() => {
-      if (isNewWallet && animationsEnabled) {
-        translateYValue.value = withTiming(0, { duration: 300 });
-        opacityValue.value = withSpring(isSelectedWallet === false ? 0.5 : 1.0, { damping: 18, stiffness: 240 });
-      }
-    }, [isNewWallet, animationsEnabled, translateYValue, opacityValue, isSelectedWallet]);
+  useEffect(() => {
+    if (isNewWallet && animationsEnabled) {
+      translateYValue.value = withTiming(0, { duration: 300 });
+      opacityValue.value = withSpring(isSelectedWallet === false ? 0.5 : 1.0, { damping: 18, stiffness: 240 });
+    }
+  }, [isNewWallet, animationsEnabled, translateYValue, opacityValue, isSelectedWallet]);
 
-    useEffect(() => {
-      if (!animationsEnabled) {
-        previousBalance.current = safeBalance;
-        return;
-      }
-
-      if (previousBalance.current !== undefined && previousBalance.current !== safeBalance) {
-        // Subtle currency-like transition on balance updates.
-        balanceOpacity.value = 0;
-        balanceTranslateY.value = 6;
-        balanceOpacity.value = withTiming(1, { duration: 180 });
-        balanceTranslateY.value = withSpring(0, { damping: 16, stiffness: 220 });
-      }
-
+  useEffect(() => {
+    if (!animationsEnabled) {
       previousBalance.current = safeBalance;
-    }, [safeBalance, animationsEnabled, balanceOpacity, balanceTranslateY]);
-
-    useEffect(() => {
-      if (isExiting && animationsEnabled) {
-        translateYValue.value = withTiming(-20, { duration: 200 });
-        opacityValue.value = withTiming(0, { duration: 200 });
-      }
-    }, [isExiting, animationsEnabled, translateYValue, opacityValue]);
-
-    const animatedCardStyle = useAnimatedStyle(() => ({
-      opacity: opacityValue.value,
-      transform: [{ scale: pressScale.value * dragScale.value }, { translateY: translateYValue.value }],
-    }));
-
-    const animatedBalanceStyle = useAnimatedStyle(() => ({
-      opacity: balanceOpacity.value,
-      transform: [{ translateY: balanceTranslateY.value }],
-    }));
-
-    let image;
-    switch (item.type) {
-      case LightningCustodianWallet.type:
-      case LightningArkWallet.type:
-        image = direction === 'rtl' ? require('../img/lnd-shape-rtl.png') : require('../img/lnd-shape.png');
-        break;
-      case MultisigHDWallet.type:
-        image = direction === 'rtl' ? require('../img/vault-shape-rtl.png') : require('../img/vault-shape.png');
-        break;
-      default:
-        image = direction === 'rtl' ? require('../img/btc-shape-rtl.png') : require('../img/btc-shape.png');
+      return;
     }
 
-    let latestTransactionText;
-    if (item.getBalance() !== 0 && item.getLatestTransactionTime() === 0) {
-      latestTransactionText = loc.wallets.pull_to_refresh;
-    } else if (walletHasPendingTransaction(item)) {
-      latestTransactionText = loc.transactions.pending;
-    } else {
-      latestTransactionText = transactionTimeToReadable(item.getLatestTransactionTime());
+    if (previousBalance.current !== undefined && previousBalance.current !== safeBalance) {
+      // Subtle currency-like transition on balance updates.
+      balanceOpacity.value = 0;
+      balanceTranslateY.value = 6;
+      balanceOpacity.value = withTiming(1, { duration: 180 });
+      balanceTranslateY.value = withSpring(0, { damping: 16, stiffness: 220 });
     }
 
-    return (
-      <Animated.View
-        style={[
-          sizeClass === SizeClass.Large || !horizontal
-            ? [iStyles.rootLargeDevice, customStyle]
-            : [iStyles.root, { width: itemWidth }, customStyle],
-          animatedCardStyle,
-        ]}
+    previousBalance.current = safeBalance;
+  }, [safeBalance, animationsEnabled, balanceOpacity, balanceTranslateY]);
+
+  useEffect(() => {
+    if (isExiting && animationsEnabled) {
+      translateYValue.value = withTiming(-20, { duration: 200 });
+      opacityValue.value = withTiming(0, { duration: 200 });
+    }
+  }, [isExiting, animationsEnabled, translateYValue, opacityValue]);
+
+  const animatedCardStyle = useAnimatedStyle(() => ({
+    opacity: opacityValue.value,
+    transform: [{ scale: pressScale.value * dragScale.value }, { translateY: translateYValue.value }],
+  }));
+
+  const animatedBalanceStyle = useAnimatedStyle(() => ({
+    opacity: balanceOpacity.value,
+    transform: [{ translateY: balanceTranslateY.value }],
+  }));
+
+  let image;
+  switch (item.type) {
+    case LightningCustodianWallet.type:
+    case LightningArkWallet.type:
+      image = direction === 'rtl' ? require('../img/lnd-shape-rtl.png') : require('../img/lnd-shape.png');
+      break;
+    case MultisigHDWallet.type:
+      image = direction === 'rtl' ? require('../img/vault-shape-rtl.png') : require('../img/vault-shape.png');
+      break;
+    default:
+      image = direction === 'rtl' ? require('../img/btc-shape-rtl.png') : require('../img/btc-shape.png');
+  }
+
+  let latestTransactionText;
+  if (item.getBalance() !== 0 && item.getLatestTransactionTime() === 0) {
+    latestTransactionText = loc.wallets.pull_to_refresh;
+  } else if (walletHasPendingTransaction(item)) {
+    latestTransactionText = loc.transactions.pending;
+  } else {
+    latestTransactionText = transactionTimeToReadable(item.getLatestTransactionTime());
+  }
+
+  return (
+    <Animated.View
+      style={[
+        sizeClass === SizeClass.Large || !horizontal
+          ? [iStyles.rootLargeDevice, customStyle]
+          : [iStyles.root, { width: itemWidth }, customStyle],
+        animatedCardStyle,
+      ]}
+    >
+      <Pressable
+        accessibilityRole="button"
+        testID={walletLabel}
+        onPressIn={onPressedIn}
+        onPressOut={onPressedOut}
+        onLongPress={() => {
+          if (handleLongPress) handleLongPress();
+        }}
+        onPress={handlePress}
+        delayHoverIn={0}
+        delayHoverOut={0}
       >
-        <Pressable
-          accessibilityRole="button"
-          testID={walletLabel}
-          onPressIn={onPressedIn}
-          onPressOut={onPressedOut}
-          onLongPress={() => {
-            if (handleLongPress) handleLongPress();
-          }}
-          onPress={handlePress}
-          delayHoverIn={0}
-          delayHoverOut={0}
+        <View
+          style={[
+            iStyles.shadowContainer,
+            isCompact && iStyles.shadowContainerCompact,
+            { backgroundColor: colors.background, shadowColor: colors.shadowColor },
+          ]}
         >
-          <View
-            style={[
-              iStyles.shadowContainer,
-              isCompact && iStyles.shadowContainerCompact,
-              { backgroundColor: colors.background, shadowColor: colors.shadowColor },
-            ]}
+          <LinearGradient
+            colors={WalletGradient.gradientsFor(item.type)}
+            style={[iStyles.grad, isCompact && iStyles.gradCompact, scaledCardStyles.grad]}
           >
-            <LinearGradient
-              colors={WalletGradient.gradientsFor(item.type)}
-              style={[iStyles.grad, isCompact && iStyles.gradCompact, scaledCardStyles.grad]}
-            >
-              <ImageBackground source={image} style={[iStyles.image, isCompact && iStyles.imageCompact]} />
-              <View style={[iStyles.gradContent, isCompact && iStyles.gradContentCompact, !isCompact && scaledCardStyles.gradContent]}>
-                {!isPlaceHolder && (
-                  <>
-                    <Text
-                      numberOfLines={1}
-                      style={[iStyles.label, isCompact && iStyles.labelCompact, scaledCardStyles.label, cardTextStyle]}
-                    >
-                      {renderHighlightedText ? renderHighlightedText(walletLabel, searchQuery || '') : walletLabel}
-                    </Text>
-                    <View
-                      style={[iStyles.balanceContainer, isCompact && iStyles.balanceContainerCompact, scaledCardStyles.balanceContainer]}
-                    >
-                      {hideBalance ? (
-                        <>
-                          <BlueSpacing10 />
-                          <BlurredBalanceView />
-                        </>
-                      ) : (
-                        <Animated.Text
-                          numberOfLines={1}
-                          adjustsFontSizeToFit
-                          minimumFontScale={0.55}
-                          key={dataRevision} // force recreation on in-place wallet updates (also fixes RTL langs like Farsi)
-                          style={[
-                            iStyles.balance,
-                            isCompact && iStyles.balanceCompact,
-                            isCompact ? scaledCardStyles.balanceCompact : scaledCardStyles.balance,
-                            cardTextStyle,
-                            animatedBalanceStyle,
-                          ]}
-                        >
-                          {`${balance} `}
-                        </Animated.Text>
-                      )}
-                    </View>
-                    <View style={scaledCardStyles.textSpacer} />
-                    <Text
-                      numberOfLines={1}
-                      adjustsFontSizeToFit
-                      minimumFontScale={0.8}
-                      style={[iStyles.latestTx, isCompact && iStyles.latestTxCompact, scaledCardStyles.latestTx, cardTextStyle]}
-                    >
-                      {loc.wallets.list_latest_transaction}
-                    </Text>
-                    <Text
-                      numberOfLines={1}
-                      adjustsFontSizeToFit
-                      minimumFontScale={0.8}
-                      style={[iStyles.latestTxTime, isCompact && iStyles.latestTxTimeCompact, scaledCardStyles.latestTxTime, cardTextStyle]}
-                    >
-                      {latestTransactionText}
-                    </Text>
-                  </>
-                )}
-              </View>
-            </LinearGradient>
-          </View>
-        </Pressable>
-      </Animated.View>
-    );
-  },
-);
+            <ImageBackground source={image} style={[iStyles.image, isCompact && iStyles.imageCompact]} />
+            <View style={[iStyles.gradContent, isCompact && iStyles.gradContentCompact, !isCompact && scaledCardStyles.gradContent]}>
+              {!isPlaceHolder && (
+                <>
+                  <Text numberOfLines={1} style={[iStyles.label, isCompact && iStyles.labelCompact, scaledCardStyles.label, cardTextStyle]}>
+                    {renderHighlightedText ? renderHighlightedText(walletLabel, searchQuery || '') : walletLabel}
+                  </Text>
+                  <View style={[iStyles.balanceContainer, isCompact && iStyles.balanceContainerCompact, scaledCardStyles.balanceContainer]}>
+                    {hideBalance ? (
+                      <>
+                        <BlueSpacing10 />
+                        <BlurredBalanceView />
+                      </>
+                    ) : (
+                      <Animated.Text
+                        numberOfLines={1}
+                        adjustsFontSizeToFit
+                        minimumFontScale={0.55}
+                        key={`${balance}`} // force recreation on balance change (RTL langs like Farsi)
+                        style={[
+                          iStyles.balance,
+                          isCompact && iStyles.balanceCompact,
+                          isCompact ? scaledCardStyles.balanceCompact : scaledCardStyles.balance,
+                          cardTextStyle,
+                          animatedBalanceStyle,
+                        ]}
+                      >
+                        {`${balance} `}
+                      </Animated.Text>
+                    )}
+                  </View>
+                  <View style={scaledCardStyles.textSpacer} />
+                  <Text
+                    numberOfLines={1}
+                    adjustsFontSizeToFit
+                    minimumFontScale={0.8}
+                    style={[iStyles.latestTx, isCompact && iStyles.latestTxCompact, scaledCardStyles.latestTx, cardTextStyle]}
+                  >
+                    {loc.wallets.list_latest_transaction}
+                  </Text>
+                  <Text
+                    numberOfLines={1}
+                    adjustsFontSizeToFit
+                    minimumFontScale={0.8}
+                    style={[iStyles.latestTxTime, isCompact && iStyles.latestTxTimeCompact, scaledCardStyles.latestTxTime, cardTextStyle]}
+                  >
+                    {latestTransactionText}
+                  </Text>
+                </>
+              )}
+            </View>
+          </LinearGradient>
+        </View>
+      </Pressable>
+    </Animated.View>
+  );
+};
 
 interface WalletsCarouselProps extends Partial<FlatListProps<any>> {
   horizontal?: boolean;
@@ -794,7 +773,6 @@ const WalletsCarousel = forwardRef<CarouselListRefType, WalletsCarouselProps>((p
           isSelectedWallet={!horizontal && selectedWallet ? selectedWallet === item.getID() : undefined}
           item={item}
           hideBalance={item.hideBalance}
-          dataRevision={getWalletCarouselItemDataRevision(item)}
           handleLongPress={handleLongPress}
           onPress={onPress}
           horizontal={horizontal}
@@ -863,7 +841,6 @@ const WalletsCarousel = forwardRef<CarouselListRefType, WalletsCarouselProps>((p
             isSelectedWallet={!horizontal && selectedWallet ? selectedWallet === item.getID() : undefined}
             item={item}
             hideBalance={item.hideBalance}
-            dataRevision={getWalletCarouselItemDataRevision(item)}
             handleLongPress={handleLongPress}
             onPress={onPress}
             searchQuery={props.searchQuery}
